## WIP: Data Pipelining

Not ready for review

## Related Issues
* #11 

## Test Plan

### Baseline Benchmark (before commencing any changes)
* `vanilla_2_procs_without_sleep`
  * 3164 rounds/sec 
* `uniform_sleep_of_10ms`
  * 22.94 rounds/sec   
* `exporter_10ms_while_others_1ms`
  * 64.31 rounds/sec   
* `importer_10ms_while_others_1ms`
  * 63.99 rounds/sec   
* `first_processor_10ms_while_others_1ms`
  * 62.35 rounds/sec   


#### Evidence
```sh
go test -benchmem -run=^$ -bench ^BenchmarkPipeline$ github.com/algorand/conduit/conduit/pipeline -timeout 5m -benchtime 10s

=== RUN   BenchmarkPipeline
BenchmarkPipeline

=== RUN   BenchmarkPipeline/vanilla_2_procs_without_sleep
BenchmarkPipeline/vanilla_2_procs_without_sleep-8                      2        5001635592 ns/op                3164 rounds/sec   56055256 B/op     838342 allocs/op

=== RUN   BenchmarkPipeline/uniform_sleep_of_10ms
BenchmarkPipeline/uniform_sleep_of_10ms-8                              2        5034083494 ns/op                22.94 rounds/sec    350900 B/op       5932 allocs/op

=== RUN   BenchmarkPipeline/exporter_10ms_while_others_1ms
BenchmarkPipeline/exporter_10ms_while_others_1ms-8                     2        5015005884 ns/op                64.31 rounds/sec   1040404 B/op      16608 allocs/op

=== RUN   BenchmarkPipeline/importer_10ms_while_others_1ms
BenchmarkPipeline/importer_10ms_while_others_1ms-8                     2        5001010800 ns/op                63.99 rounds/sec   1034500 B/op      16483 allocs/op

=== RUN   BenchmarkPipeline/first_processor_10ms_while_others_1ms
BenchmarkPipeline/first_processor_10ms_while_others_1ms-8              2        5011949026 ns/op                62.35 rounds/sec   1035548 B/op      16084 allocs/op
```

### After pipelining the pipeline - the original approach

#### statistics

* `vanilla_2_procs_without_sleep`
  * 3969 rounds/sec (25% faster)
* `uniform_sleep_of_10ms`
  * 81.59 rounds/sec (256% faster)
* `exporter_10ms_while_others_1ms`
  * 81.23 rounds/sec (26% faster)
* `importer_10ms_while_others_1ms`
  * 91.97 rounds/sec (44% faster)
* `first_processor_10ms_while_others_1ms`
  * 90.84 rounds/sec (46% faster)

### After the signals approach (with `concurrentRounds = 2`)

### statistics

* `vanilla_2_procs_without_sleep`
  * 3257 rounds/sec (3% faster)
* `uniform_sleep_of_10ms`
  * 43.25 rounds/sec (89% faster)
* `exporter_10ms_while_others_1ms`
  * 82.32 rounds/sec (28% faster)
* `importer_10ms_while_others_1ms`
  * 119.2 rounds/sec (86% faster)
* `first_processor_10ms_while_others_1ms`
  * 120.0 rounds/sec (92% faster)

### After the signals approach (with `concurrentRounds = 3`)

### statistics

* `vanilla_2_procs_without_sleep`
  * 3828 rounds/sec (21% faster)
* `uniform_sleep_of_10ms`
  * 65.27 rounds/sec (185% faster)
* `exporter_10ms_while_others_1ms`
  * 81.17 rounds/sec (26% faster)
* `importer_10ms_while_others_1ms`
  * 187.4 rounds/sec (193% faster)
* `first_processor_10ms_while_others_1ms`
  * 193.4 rounds/sec (210% faster)

### After the signals approach (with `concurrentRounds = 10`)

### statistics

* `vanilla_2_procs_without_sleep`
  * 3855 rounds/sec (22% faster)
* `uniform_sleep_of_10ms`
  *  82.40 rounds/sec (259% faster)
* `exporter_10ms_while_others_1ms`
  * 81.14 rounds/sec (26% faster)
* `importer_10ms_while_others_1ms`
  * 627.1 rounds/sec (880% faster)
* `first_processor_10ms_while_others_1ms`
  * 541.2 rounds/sec (768% faster)

## Approach outlined by Will after design discussion

* `vanilla_2_procs_without_sleep`
  * 3038 rounds/sec (4% _SLOWER_)
* `uniform_sleep_of_10ms`
  *  83.20 rounds/sec (263% faster)
* `exporter_10ms_while_others_1ms`
  * 83.08 rounds/sec (29% faster)
* `importer_10ms_while_others_1ms`
  * 92.44 rounds/sec (44% faster)
* `first_processor_10ms_while_others_1ms`
  * 92.51 rounds/sec (48% faster)